+++ src/db/schemas/Card.js
@@ -17,9 +17,9 @@ const CardSchema = new Schema({
         min: 1,
         max: 3,
     },
-    timeToGuess: {
+    playingTime: {
         type: Number,
-        min: 1,
+        min: 0,
     },
     description: { type: String },
     imageURL: { type: String },