+++ src/db/schemas/Game.js
@@ -7,6 +7,19 @@ const GameSummarySchema = require("./GameSummary");
 const GameOptionsSchema = require("./GameOptions");
 const GameHistorySchema = require("./GameHistory");
 const { getGameStatuses, getGameDefaultOptions, getGameCardById } = require("../../helpers/functions/Game");
+const { getAuthStrategyFromReq } = require("../../helpers/functions/Passport");
+const { generateError } = require("../../helpers/functions/Error");
+
+const AnonymousUserSchema = new Schema({
+    _id: {
+        type: String,
+        required: true,
+    },
+}, {
+    _id: false,
+    timestamps: false,
+    versionKey: false,
+});
 
 const GameSchema = new Schema({
     players: {
@@ -17,6 +30,7 @@ const GameSchema = new Schema({
         type: [CardSchema],
         required: true,
     },
+    anonymousUser: { type: AnonymousUserSchema },
     status: {
         type: String,
         required: true,
@@ -88,6 +102,15 @@ GameSchema.virtual("isOver").get(isOver);
 GameSchema.virtual("firstQueue").get(getFirstQueue);
 GameSchema.virtual("nextSpeaker").get(getNextSpeaker);
 
+function checkBelongsToUserFromReq(req) {
+    if (getAuthStrategyFromReq(req) === "JWT") {
+        const { mode, _id } = req.user;
+        if (mode === "anonymous" && (!this.anonymousUser || _id.toString() !== this.anonymousUser?._id.toString())) {
+            throw generateError("GAME_DOESNT_BELONG_TO_USER", `Game with id ${this._id} doesn't belong to user with id "${_id}".`);
+        }
+    }
+}
+
 function getCardById(id) {
     return getGameCardById(this, id);
 }
@@ -143,7 +166,7 @@ function pushSummaryRound() {
 function resetCardsForNewRound() {
     this.cards.forEach(card => {
         card.set("status", "to-guess");
-        card.set("timeToGuess", undefined);
+        card.set("playingTime", undefined);
     });
 }
 
@@ -191,6 +214,7 @@ function setFinalSummary() {
     this.set("summary.winners", winners);
 }
 
+GameSchema.methods.checkBelongsToUserFromReq = checkBelongsToUserFromReq;
 GameSchema.methods.getCardById = getCardById;
 GameSchema.methods.getPlayersByTeam = getPlayersByTeam;
 GameSchema.methods.rollQueue = rollQueue;