+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # ⏳ LIST OF CHANGES FOR TIME'S UP API
 
+## 0.1.2 (2021-12-09)
+
+### 🚀 New features
+
+* [#17](https://github.com/pIay-it/times-up-api/issues/17) - JWT authentication for anonymous users.
+* [#18](https://github.com/pIay-it/times-up-api/issues/18) - Decimal seconds for `playingTime` card's field.
+* [#19](https://github.com/pIay-it/times-up-api/issues/19) - Change `timeToGuess` for `playingTime` for game's cards.
+* [#20](https://github.com/pIay-it/times-up-api/issues/20) - Prevent to play card with `to-guess` status.
+
+### 📦 Packages
+
+* `jsonwebtoken` installed with version `8.5.1`.
+* `passport-jwt` installed with version `4.0.0`.
+* `passport-local` installed with version `1.0.0`.
+* `uniqid` installed with version `5.4.0`.
+* `@sentry/node` updated to version `6.16.0`.
+* `@sentry/tracing` updated to version `6.16.0`.
+* `eslint` updated to version `8.4.1`.
+* `mongoose` updated to version `6.1.0`.
+
+---
+
 ## 0.1.1 (2021-11-30)
 
 ### 🚀 New features
@@ -22,9 +44,9 @@
 
 ### 📦 Packages
 
-* `apidoc` updated to version `0.50.2`
-* `eslint` updated to version `8.3.0`
-* `mongoose` updated to version `6.0.14`
+* `apidoc` updated to version `0.50.2`.
+* `eslint` updated to version `8.3.0`.
+* `mongoose` updated to version `6.0.14`.
 
 ---